fix: do not skip protocol messages (#2578)
* fix: do not skip protocol messages

* add skip get request if operation completed
