+++ src/commands/protocols/common/protocol-message-command.js
@@ -1,9 +1,5 @@
 import Command from '../../command.js';
-import {
-    NETWORK_MESSAGE_TYPES,
-    OPERATION_REQUEST_STATUS,
-    OPERATION_STATUS,
-} from '../../../constants/constants.js';
+import { NETWORK_MESSAGE_TYPES, OPERATION_REQUEST_STATUS } from '../../../constants/constants.js';
 
 class ProtocolMessageCommand extends Command {
     constructor(ctx) {
@@ -20,19 +16,8 @@ class ProtocolMessageCommand extends Command {
         return this.sendProtocolMessage(command, message, messageType);
     }
 
-    async shouldSendMessage(command) {
-        const { operationId } = command.data;
-
-        const { status } = await this.operationService.getOperationStatus(operationId);
-
-        if (status === OPERATION_STATUS.IN_PROGRESS) {
-            return true;
-        }
-        this.logger.trace(
-            `${command.name} skipped for operationId: ${operationId} with status ${status}`,
-        );
-
-        return false;
+    async shouldSendMessage() {
+        return true;
     }
 
     async prepareMessage() {
