+++ src/commands/protocols/get/sender/v1.0.0/v1-0-0-get-request-command.js
@@ -3,6 +3,7 @@ import {
     NETWORK_MESSAGE_TIMEOUT_MILLS,
     ERROR_TYPE,
     OPERATION_REQUEST_STATUS,
+    OPERATION_STATUS,
 } from '../../../../../constants/constants.js';
 
 class GetRequestCommand extends ProtocolRequestCommand {
@@ -12,6 +13,21 @@ class GetRequestCommand extends ProtocolRequestCommand {
         this.validationService = ctx.validationService;
 
         this.errorType = ERROR_TYPE.GET.GET_REQUEST_ERROR;
+    }
+
+    async shouldSendMessage(command) {
+        const { operationId } = command.data;
+
+        const { status } = await this.operationService.getOperationStatus(operationId);
+
+        if (status === OPERATION_STATUS.IN_PROGRESS) {
+            return true;
+        }
+        this.logger.trace(
+            `${command.name} skipped for operationId: ${operationId} with status ${status}`,
+        );
+
+        return false;
     }
 
     async prepareMessage(command) {
